Split Android navigation packages
