+++ apps/android/app/src/androidTest/java/com/flashcardsopensourceapp/app/GuestSignInAfterReviewPromptDialogTest.kt
@@ -13,7 +13,7 @@ import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import androidx.navigation.compose.rememberNavController
 import androidx.test.ext.junit.runners.AndroidJUnit4
-import com.flashcardsopensourceapp.app.navigation.SettingsAccountSignInEmailDestination
+import com.flashcardsopensourceapp.app.navigation.settings.SettingsAccountSignInEmailDestination
 import com.flashcardsopensourceapp.core.ui.theme.FlashcardsTheme
 import com.flashcardsopensourceapp.feature.settings.cloud.CloudSignInEmailRoute
 import com.flashcardsopensourceapp.feature.settings.cloud.CloudSignInUiState
