+++ apps/android/app/src/main/java/com/flashcardsopensourceapp/app/navigation/progress/ProgressNavGraph.kt
@@ -1,4 +1,4 @@
-package com.flashcardsopensourceapp.app.navigation
+package com.flashcardsopensourceapp.app.navigation.progress
 
 import androidx.compose.runtime.getValue
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
@@ -6,6 +6,7 @@ import androidx.lifecycle.viewmodel.compose.viewModel
 import androidx.navigation.NavGraphBuilder
 import androidx.navigation.compose.composable
 import com.flashcardsopensourceapp.app.di.AppGraph
+import com.flashcardsopensourceapp.app.navigation.ProgressDestination
 import com.flashcardsopensourceapp.feature.progress.ProgressRoute
 import com.flashcardsopensourceapp.feature.progress.ProgressViewModel
 import com.flashcardsopensourceapp.feature.progress.createProgressViewModelFactory
