+++ apps/android/app/src/main/java/com/flashcardsopensourceapp/app/navigation/review/ReviewDestinations.kt
@@ -1,4 +1,4 @@
-package com.flashcardsopensourceapp.app.navigation
+package com.flashcardsopensourceapp.app.navigation.review
 
 data object ReviewPreviewDestination {
     const val route: String = "review/preview"
